Update of RCUR_4_debug.h and added main.cpp file
