+++ src/RIDDecomposition.cpp
