+++ src/RIDDecomposition.h
