+++ src/RQBDecomposition.h
